krausening-python/pyproject.toml
[tool.poetry]
name = "krausening"
version = "11.dev"
description = "Python implementation of Krausening"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.9"
javaproperties = "^0.8.1"
pycryptodome = "^3.14.1"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"
nose = "^1.3.7"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
